code polishing

+++ src/main/java/com/mirotic/demorestapi/events/EventRepository.java
@@ -2,5 +2,5 @@ package com.mirotic.demorestapi.events;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
 }
